Extract RawResultCell to separate file
